Fixups for rename of drivers/cpu to drivers/core.
